LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-optgroup-arrow-keys.optional-expected.txt



FAIL Keyboard navigation forwards and backwards should visit each option with optgroups. assert_equals: appearance:base-select must be supported to run this test. expected "base-select" but got "auto"
